chore(firebase): refactoring, moved getBoards to firebase utils
