+++ next-tavla/src/Shared/utils/firebase.ts
@@ -10,6 +10,7 @@ import {
     getFirestore,
 } from '@firebase/firestore/lite'
 import { FIREBASE_CLIENT_CONFIG } from 'assets/env'
+import { upgradeSettings } from './converters'
 
 const app = initializeClientApp()
 export const auth = getAuth(app)
@@ -43,4 +44,33 @@ export async function setBoardSettings(boardId: string, settings: TSettings) {
 export async function addBoardSettings(settings: TSettings) {
     const firestore = safeGetFirestore()
     return await addDoc(collection(firestore, 'settings-v2'), settings)
+}
+
+export async function getBoards() {
+    //when login is fixed:
+    // 1. take user id as params
+    // 2. create getBoards function in firebase.ts that gets all boards for a user based on user id
+    // 3. use getBoards function instead of getBoardSettings to get all boards for a user
+    // 4. remove ids const
+    const ids = [
+        'Malre1Dx5zLE086AzpFH',
+        'oMgfeCRUZ4sfD8xXXG8M',
+        '8EyKHHP5Ie2HisOO2ilt',
+    ]
+
+    const settingsPromises = ids.map(async (id: string) => {
+        const settings = await getBoardSettings(id)
+        return { id, settings }
+    })
+
+    const allSettingsWithIds = await Promise.all(settingsPromises)
+
+    const convertedSettings = allSettingsWithIds.map(({ id, settings }) => {
+        return {
+            id,
+            settings: settings ? upgradeSettings(settings) : null,
+        }
+    })
+
+    return convertedSettings
 }
